backend/requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
Pillow
django-cors-headers
